requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
albumentations==1.3.1
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
axial-positional-embedding==0.2.1
backcall==0.2.0
beautifulsoup4==4.12.2
benchpots==0.1.1
blessed==1.20.0
boto3==1.34.112
botocore==1.34.112
cachetools==5.3.2
captum==0.7.0
certifi==2022.12.7
charset-normalizer==3.1.0
clarabel==0.9.0
click==8.1.3
cmake==3.26.1
CoLT5-attention==0.10.20
contextlib2==21.6.0
contourpy==1.0.7
croniter==1.3.10
cvxopt==1.3.2
cvxpy==1.5.2
cycler==0.11.0
dateutils==0.6.12
decorator==5.1.1
deepdiff==6.3.0
dnspython==2.3.0
docker-pycreds==0.4.0
docstring-parser==0.15
ecos==2.0.14
einops==0.7.0
email-validator==1.3.1
executing==1.2.0
fancyimpute==0.7.0
fastapi==0.88.0
filelock==3.11.0
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.12.2
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.24.0
google-auth-oauthlib==1.1.0
grpcio==1.59.3
h5py==3.11.0
httptools==0.5.0
huggingface-hub==0.20.3
hydra-core==1.3.2
imageio==2.31.1
imbalanced-learn==0.12.2
imblearn==0.0
import-ipynb==0.1.4
importlib-resources==5.12.0
iniconfig==2.0.0
inquirer==3.1.3
ipdb==0.13.13
ipython==8.14.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonargparse==4.20.1
kiwisolver==1.4.4
knnimpute==0.1.0
latextable==1.0.1
lazy_loader==0.3
lightning==2.1.4
lightning-cloud==0.5.32
lightning-utilities==0.8.0
linear-attention-transformer==0.19.1
linformer==0.2.3
lit==16.0.0
llvmlite==0.43.0
local-attention==1.9.1
Markdown==3.5.1
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-collections==0.1.1
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.4
networkx==3.1
nltk==3.8.1
nose==1.3.7
numba==0.60.0
numpy==1.24.3
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.7.0.72
opencv-python-headless==4.8.0.74
ordered-set==4.1.0
orjson==3.8.9
osqp==0.6.7.post0
packaging==23.0
pandas==1.5.3
parso==0.8.3
pathtools==0.1.2
pickleshare==0.7.5
Pillow==9.5.0
pluggy==1.5.0
ply==3.11
POT==0.9.3
product-key-memory==0.2.10
prompt-toolkit==3.0.39
protobuf==3.20.3
psutil==5.9.5
pure-eval==0.2.2
pyarrow==16.1.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==1.10.7
pydicom==2.3.1
Pygments==2.14.0
pygrinder==0.6.1
PyJWT==2.6.0
pynndescent==0.5.13
pyparsing==3.0.9
pypots==0.5
PyQt5==5.15.10
pyrootutils==1.0.4
pytest==8.2.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-editor==1.0.4
python-multipart==0.0.6
pytorch-lightning==2.0.5
pytorch-pretrained-bert==0.6.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
qdldl==0.1.7.post4
qudida==0.0.4
ray==2.46.0
readchar==4.0.5
regex==2023.3.23
requests==2.28.2
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==13.3.3
rsa==4.9
s3transfer==0.10.1
safetensors==0.4.2
SciencePlots==2.1.1
scikit-image==0.21.0
scikit-learn==1.3.0
scs==3.2.5
seaborn==0.13.2
sentry-sdk==1.19.1
setproctitle==1.3.2
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4
stack-data==0.6.2
starlette==0.22.0
starsessions==1.3.0
sympy==1.11.1
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorboardX==2.6
texttable==1.7.0
threadpoolctl==3.2.0
tifffile==2023.7.18
timm==0.9.16
tokenizers==0.15.1
tomli==2.0.1
torch-tb-profiler==0.4.3
torchdiffeq==0.2.3
torchmetrics==0.11.4
tqdm==4.65.0
traitlets==5.9.0
tsdb==0.5
typeshed-client==2.2.0
typing_extensions==4.7.1
tzdata==2023.3
ujson==5.7.0
umap-learn==0.5.6
urllib3==1.26.15
uvicorn==0.21.1
wandb==0.16.2
watchfiles==0.19.0
webencodings==0.5.1
websocket-client==1.5.1
websockets==11.0.1
Werkzeug==3.0.1
xgboost==2.0.3
yarl==1.8.2
transformers==4.36.2